# api/openapi.yaml
openapi: 3.1.0
info:
  title: Aperture API
  version: 0.0.0
  description: |
    REST API for managing Aperture tenants, browser sessions, snapshots, API tokens, and tenant events.

    ## Authentication

    Send `Authorization: Bearer <token>` or authenticate through an account browser session established by OIDC, passkey, or password. Bearer credentials take precedence when both are present. The health check is public. `system:admin` grants every scope. Tenant API tokens may additionally restrict access to explicit session and snapshot IDs. Raw tokens are returned only on creation or rotation.

    ## Tenant selection

    Tenant API tokens use their bound tenant. System administrators and account sessions with multiple memberships select a tenant with `X-Aperture-Tenant-Id` for tenant-scoped operations. Aperture selects the only membership automatically when an account has one.

    ## Pagination

    Lists are newest-first. Pass `meta.nextCursor` as `cursor` while `meta.hasMore` is `true`. Page size defaults to 50 and is capped at 100.

    ## Errors

    Errors use `{ "error": { "code": "...", "message": "..." } }`. Each operation enumerates its possible stable codes; `ErrorCode` is the complete catalog. Request bodies are limited to 1 MiB.
servers:
  - url: /
security:
  - bearerAuth: []
  - webSession: []
...
paths:
...
  /api/sessions:
    parameters:
      - {$ref: '#/components/parameters/SelectedTenantId'}
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List browser sessions
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:read]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, session_service_unavailable]
      description: Newest-first sessions. A resource-restricted token receives only granted sessions; filtering occurs before pagination. Credentials are included only while available and must be treated as secrets.
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/IncludeDeleted'}
        - name: status
          in: query
          description: Exact session lifecycle status.
          schema: {$ref: '#/components/schemas/SessionStatus'}
        - {$ref: '#/components/parameters/TagKey'}
        - {$ref: '#/components/parameters/TagValue'}
        - {$ref: '#/components/parameters/TagOperator'}
      responses:
        '200': {description: Session page, content: {application/json: {schema: {$ref: '#/components/schemas/SessionPage'}}}}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a browser session
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:write]
        conditionalScopes:
          - scope: snapshots:read
            condition:
              source: requestBody
              property: baseSnapshotName
              operator: non_empty
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [invalid_request_body, validation_failed, resource_access_denied, base_snapshot_not_found, base_snapshot_deleted, session_service_unavailable, overlay_mount_failed, browser_start_failed]
      description: Starts a browser with a retained overlay. A resource-restricted token may start a blank session or use a granted base snapshot. The new session is not added to its resource allowlist, but the returned session token remains usable. Failed startup may leave a reopenable `failed` session.
      requestBody: {$ref: '#/components/requestBodies/CreateSession'}
      responses:
        '201': {description: Created session, content: {application/json: {schema: {$ref: '#/components/schemas/CreateSessionResult'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /api/sessions/{sessionId}:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
      - {$ref: '#/components/parameters/SelectedTenantId'}
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a browser session
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:read]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, resource_access_denied, session_not_found, session_expired, session_service_unavailable]
      description: Session metadata and current credentials when available.
      responses:
        '200': {description: Browser session, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a browser session
      x-aperture-auth:
        authorityTypes: [system_admin, tenant]
        requiredScopes: [sessions:write]
        tenantResolution: token_tenant_or_required_header
      x-aperture-errors: [validation_failed, resource_access_denied, session_not_found, session_expired, session_service_unavailable, overlay_mount_failed, browser_control_failed]
      description: Stops and marks the session `deleted`. The retained overlay remains reopenable until `expiresAt`.
      responses:
        '200': {description: Deleted session, content: {application/json: {schema: {$ref: '#/components/schemas/SessionMutation'}}}}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Aperture API token
      description: |
        Aperture API token sent as `Authorization: Bearer <token>`. API tokens normally start with `apt_`. Session access tokens use a separate `aps_` format and are not accepted as API authentication.
    webSession:
      type: apiKey
      in: cookie
      name: aperture_session
      description: Server-side browser session established by OIDC, passkey, or password login.
  parameters:
    Limit:
      name: limit
      in: query
      description: Requested page size. Omitted or `0` uses 50. Values above 100 are accepted but clamped to 100.
      schema:
        type: integer
        minimum: 0
        default: 50
      example: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque `nextCursor` from the previous page. Reuse it only with the same operation and filters.
      schema: {type: string}
      example: eyJjcmVhdGVkQXQiOiIyMDI2LTA3LTE3VDEyOjAwOjAwWiIsImlkIjoiMDE5ZjZjZjAtMDAwMC03MDAwLTgwMDAtMDAwMDAwMDAwMDAwIn0
    IncludeDeleted:
      name: includeDeleted
      in: query
      description: Include soft-deleted sessions when `true`. The default is `false`.
      schema: {type: boolean, default: false}
...
    SelectedTenantId:
      name: X-Aperture-Tenant-Id
      in: header
      description: |
        Tenant selected for a tenant-scoped operation. System administrators and account sessions may provide this header. A tenant API token uses its bound tenant and may omit the header; selecting a different tenant is forbidden.
      schema: {$ref: '#/components/schemas/UUIDv7'}
      example: 019f6cf0-0000-7000-8000-000000000001
...
    SessionId:
      name: sessionId
      in: path
      required: true
      description: UUIDv7 identifier of the browser session.
      schema: {$ref: '#/components/schemas/UUIDv7'}
...
    TagKey:
      name: tagKey
      in: query
      description: |
        Repeatable tag key. Each entry pairs by position with `tagValue` and, when supplied, `tagOperator`. Multiple predicates are combined with logical AND.
      schema: {type: array, items: {type: string, minLength: 1}}
      style: form
      explode: true
      example: [environment, region]
    TagValue:
      name: tagValue
      in: query
      description: |
        Repeatable tag value paired with `tagKey`. `eq` and `neq` accept one value. `in` and `not_in` accept comma-separated values in the matching entry.
      schema: {type: array, items: {type: string, minLength: 1}}
      style: form
      explode: true
      example: [production, 'eu-west-1,eu-central-1']
    TagOperator:
      name: tagOperator
      in: query
      description: |
        Optional operator paired by position with each tag predicate. Omit the entire parameter to use `eq` for every predicate. If supplied, its entry count must match `tagKey`.
      schema:
        type: array
        items:
          type: string
          enum: [eq, neq, in, not_in]
      style: form
      explode: true
      example: [eq, in]
  requestBodies:
...
    CreateSession:
      required: true
      description: Browser session configuration. The JSON body must not exceed 1 MiB.
      content: {application/json: {schema: {$ref: '#/components/schemas/CreateSessionInput'}}}
...
  responses:
    Error:
      description: API error. Branch on the HTTP status and stable `error.code`.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            invalidRequestBody:
              summary: 400 · invalid JSON body
              value: {error: {code: invalid_request_body, message: invalid request body}}
            validationFailed:
              summary: 400 · invalid field or parameter
              value: {error: {code: validation_failed, message: 'validation error\nname is required'}}
            tenantSelectionRequired:
              summary: 400 · tenant header required
              value: {error: {code: tenant_selection_required, message: tenant selection required}}
            authenticationRequired:
              summary: 401 · bearer token missing
              value: {error: {code: authentication_required, message: authentication required}}
            authenticationTokenExpired:
              summary: 401 · bearer token expired
              value: {error: {code: authentication_token_expired, message: authentication token expired}}
            insufficientScope:
              summary: 403 · required scope missing
              value: {error: {code: insufficient_scope, message: insufficient scope}}
            resourceAccessDenied:
              summary: 403 · resource is not granted
              value: {error: {code: resource_access_denied, message: resource access denied}}
            tenantSelectionNotPermitted:
              summary: 403 · tenant token selected another tenant
              value: {error: {code: tenant_selection_not_permitted, message: tenant selection not permitted}}
            sessionNotFound:
              summary: 404 · session missing or foreign
              value: {error: {code: session_not_found, message: session not found}}
            sessionInvalidState:
              summary: 409 · operation invalid for current state
              value: {error: {code: session_invalid_state, message: session is in an invalid state for this operation}}
            sessionExpired:
              summary: 410 · retained session expired
              value: {error: {code: session_expired, message: session expired}}
            internalError:
              summary: 500 · unexpected server failure
              value: {error: {code: internal_error, message: internal server error}}
            browserStartFailed:
              summary: 502 · browser failed to start
              value: {error: {code: browser_start_failed, message: browser failed to start}}
  schemas:
...
    SessionStatus:
      type: string
      description: |
        Session lifecycle state:

        - `creating`: browser startup is in progress.
        - `running`: browser and routing are available.
        - `suspended`: browser is intentionally stopped while retained state remains available.
        - `deleted`: session is soft-deleted and may be reopened while retained.
        - `expired`: retention elapsed and the session can no longer be used.
        - `failed`: startup or reopen failed; retained state may still be reopenable.
      enum: [creating, running, suspended, deleted, expired, failed]
      example: running
...
    SessionPage:
      type: object
      description: Cursor-paginated browser sessions.
      required: [data, meta]
      properties: {data: {type: array, items: {$ref: '#/components/schemas/Session'}}, meta: {$ref: '#/components/schemas/PageMeta'}}
...
    CreateSessionResult:
      type: object
      description: Newly created session. Active connection data is nested in the session.
      required: [session]
      properties:
        session: {$ref: '#/components/schemas/Session'}
...